requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
openai==1.3.0
python-docx==1.1.0
pdfplumber==0.10.3